Warning-free clang build
1. ✍️ Subdir test/ is excluded from build by default. To include test/, configure cmake with -DBUILD_TEST=1
2. ✍️ `bux::C_EzArgs::parse()` accepts `argc` parameter both in type **size_t** or **int**
3. ✍️ `Intervals.h` rehauled to use C++20 concepts to smartly reduce lines and eliminate clang warnings altogether.
4. ✍️ Eliminate clang warnings here & there.
5. 🆕 test/hrtn to demangle external symbols freely.

+++ src/EZArgs.cpp
@@ -142,7 +142,7 @@ C_ErrorOrIndex C_EZArgs::parse(size_t argc, const char *argv[]) const
         {
             const auto flag = arg + 2;
             const auto eqsign = strchr(flag, '=');
-            const auto flag_name = eqsign? std::string_view(flag, eqsign-flag): std::string_view(flag);
+            const auto flag_name = eqsign? std::string_view(flag, size_t(eqsign-flag)): std::string_view(flag);
             for (auto cur = this; cur; cur = cur->m_owner)
                 for (auto &def: cur->m_flags)
                     if (def.m_name == flag_name)
